feat: introduce ExpressionParser and move code into it
